# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sla/policies": {
      "get": {
        "description": "List SLA policies in calendar minutes",
        "operationId": "SlaController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlaPolicy"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or workflow transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid bearer token; inactive user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or ticket ownership does not permit this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate resource or operation conflicts with current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List SLA policies in calendar minutes",
        "tags": [
          "SLA"
        ]
      }
    },
    "/api/sla/policies/{priority}": {
      "patch": {
        "description": "Update a priority SLA policy (admin)",
        "operationId": "SlaController_update",
        "parameters": [
          {
            "name": "priority",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "URGENT"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSlaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlaPolicy"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or workflow transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid bearer token; inactive user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or ticket ownership does not permit this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate resource or operation conflicts with current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a priority SLA policy (admin)",
        "tags": [
          "SLA"
        ]
      }
    },
    "/api/sla/escalations/run": {
      "post": {
        "description": "Run the idempotent escalation sweep now (admin)",
        "operationId": "SlaController_sweep",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or workflow transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid bearer token; inactive user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or ticket ownership does not permit this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate resource or operation conflicts with current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run the idempotent escalation sweep now (admin)",
        "tags": [
          "SLA"
        ]
      }
    },
...
  },
  "info": {
    "title": "Student Support & Ticket Management",
    "description": "College Student Support Operations API for managing student tickets, assignments, SLAs, comments, resolutions, and audit history. Supports role-based access for ADMIN, STAFF, and STUDENT users, JWT authentication with Bearer tokens, UTC timestamps, and calendar-minute SLA calculations with consistent API error responses.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid ticket transition"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "SlaPolicy": {
        "type": "object",
        "properties": {
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "responseMinutes": {
            "type": "number",
            "example": 240
          },
          "resolutionMinutes": {
            "type": "number",
            "example": 1440
          }
        },
        "required": [
          "priority",
          "responseMinutes",
          "resolutionMinutes"
        ]
      },
      "UpdateSlaDto": {
        "type": "object",
        "properties": {
          "responseMinutes": {
            "type": "number",
            "example": 120,
            "minimum": 1,
            "maximum": 43200,
            "description": "Response deadline in calendar minutes; applies to new/reopened/reprioritized tickets."
          },
          "resolutionMinutes": {
            "type": "number",
            "example": 480,
            "minimum": 1,
            "maximum": 43200,
            "description": "Must be at least responseMinutes. Existing tickets retain their SLA snapshot."
          }
        },
        "required": [
          "responseMinutes",
          "resolutionMinutes"
        ]
      },
      "SweepResult": {
        "type": "object",
        "properties": {
          "escalated": {
            "type": "number",
            "description": "Number of new escalation events recorded."
          }
        },
        "required": [
          "escalated"
        ]
      },
...
    }
  }
}